요청 및 응답을 위한 각각의 모델 생성
- 포켓몬 정보
- base : 도감 리스트에 보여질 기본 모습관련 데이터
- dex : 지역별 도감 정보
- text entries : 도감 설명
