+++ PokeDex_Developer/Model/Pokemons/Base.swift
@@ -1,0 +1,14 @@
+//
+//  Base.swift
+//  PokeDex_Developer
+//
+//  Created by 유영웅 on 5/1/24.
+//
+
+import Foundation
+
+struct Base :Codable{
+    var image : String
+    var types : [String]
+    
+}
