+++ PokeDex_Developer/Model/Pokemons/Dex.swift
@@ -1,0 +1,13 @@
+//
+//  Dex.swift
+//  PokeDex_Developer
+//
+//  Created by 유영웅 on 5/1/24.
+//
+
+import Foundation
+
+struct Dex :Codable{
+    var num : Int
+    var region : String
+}
