+++ PokeDex_Developer/Model/Pokemons/TextEntries.swift
@@ -1,0 +1,13 @@
+//
+//  TextEntries.swift
+//  PokeDex_Developer
+//
+//  Created by 유영웅 on 5/1/24.
+//
+
+import Foundation
+
+struct TextEntries:Codable{
+    var text:[String]
+    var version : [String]
+}
