+++ PokeDex_Developer/Model/Response.swift
@@ -6,3 +6,9 @@
 //
 
 import Foundation
+
+struct Response: Codable {
+    let status: Int
+    let data: [Int]
+    let message : String
+}
